Moved get_excerpts to df_utils and fixes #141
