+++ mdtk/df_utils.py
@@ -3,6 +3,7 @@
 import itertools
 
 import pandas as pd
+import numpy as np
 
 
 NOTE_DF_SORT_ORDER = ['onset', 'track', 'pitch', 'dur']
@@ -99,3 +100,73 @@ def remove_pitch_overlaps(df):
     df = df.reset_index(drop=True)
 
     return df
+
+
+def get_random_excerpt(note_df, min_notes=10, excerpt_length=5000,
+                       first_onset_range=(0, 200), iterations=10):
+    """
+    Take a random excerpt from the given note_df, using np.random. The excerpt
+    is created as follows:
+
+    1. Pick a note at random from the input df, excluding the last `min_notes`
+       notes.
+    2. Take all notes which onset within `excerpt_length` ms of that note.
+    3. If the excerpt does not contain at least `min_notes` notes, repeat steps
+       1 and 2 until you have drawn `iterations` invalid excerpts. In that
+       case, return None.
+    4. If you have a valid excerpt, shift its notes so that the first note's
+       onset is at time 0, then add a random number within `first_onset_range`
+       to each onset.
+    5. Return the resulting excerpt.
+
+    Parameters
+    ----------
+    note_df : pd.DataFrame
+        The input note_df, from which we want a random excerpt.
+
+    min_notes : int
+        The minimum number of notes that must be contained in a valid excerpt.
+
+    excerpt_length : int
+        The length of the resulting excerpt, in ms. All notes which onset
+        within this amount of time after a randomly chosen note will be
+        included in the returned excerpt.
+
+    first_onset_range : tuple(int, int)
+        The range from which to draw a random number to add to the first note's
+        onset (in ms), rather than having the chosen excerpt begin at time 0.
+
+    iterations : int
+        How many times to try to obtain a valid excerpt before giving up and
+        returning None.
+
+    Returns
+    -------
+    excerpt : pd.DataFrame
+        A random excerpt from the given note_df. None if no valid excerpt was
+        found within `iterations` attempts.
+    """
+    if len(note_df) < min_notes or iterations == 0:
+        return None
+
+    for _ in range(iterations):
+        note_index = np.random.choice(list(note_df.index.values)
+                                      [:-min_notes])
+        first_onset = note_df.loc[note_index]['onset']
+        excerpt = pd.DataFrame(
+            note_df.loc[note_df['onset'].between(
+                first_onset, first_onset + excerpt_length)])
+
+        # Check for validity of excerpt
+        if len(excerpt) < min_notes:
+            excerpt = None
+        else:
+            break
+
+    if excerpt is None:
+        return None
+
+    onset_shift = np.random.randint(first_onset_range[0], first_onset_range[1])
+    excerpt['onset'] += onset_shift - first_onset
+    excerpt = excerpt.reset_index(drop=True)
+    return excerpt
